Make Settings a shared singleton
Using Settings like a singleton, and refactoring load and save into this class improves seperation concerns.
Using files for classes in a seperated dir improves conflict handling in git and helps indexing the classes.
Added unittest for the new Settings class using pytest framework. pytest chosen from previous experience.
use pytest in root folder to run tests.
With this change, foundations for improved layout of model classes and testability is established.

Saving and reusing the last used comm-port in the settings was the catalyst for this change, and will follow soon :D