feat: Enhance Stratagem Filtering Protocols
For Super Earth!

This commit enhances our stratagem filtering technology to allow for more efficient and precise selection of stratagems in the heat of battle.

The filter now ignores special characters when they are not present in the filter string, but enforces them when they are. This means you can find your 'W.A.S.P.' launcher by just typing 'wasp', but if you need to be specific with 'gl-', it won't show you any of that Eagle business.

Includes new testing protocols to ensure our new filtering protocols are up to Super Earth standards.

For Liberty!